Factor out MainWindow

+++ client-slint/src/constants.rs
@@ -1,5 +1,5 @@
 pub const WINDOW_TITLE: &str = "Walpurgis";
-pub const WINDOW_SCALE: i32 = 2;
+pub const WINDOW_SCALE: f32 = 2.0;
 pub const WINDOW_WIDTH: i32 = 160;
 pub const WINDOW_HEIGHT: i32 = 144;
 pub const TILE_WIDTH: i32 = 8;
